gh-2038 GetFromEndpoint (#2039)
* gh-2038 - changed GetAsElementsFromEndpoint to GetFromEndpoint

* gh-2038 - Added operation declarations for GetFromEndpoint operation

* Apply suggestions from code review

Updated copyright header dates
