+++ core/store/src/main/resources/GetFromEndpointOperationDeclarations.json
@@ -1,0 +1,10 @@
+{
+  "operations": [
+    {
+      "operation": "uk.gov.gchq.gaffer.operation.impl.get.GetFromEndpoint",
+      "handler": {
+        "class": "uk.gov.gchq.gaffer.store.operation.handler.GetFromEndpointHandler"
+      }
+    }
+  ]
+}
