+++ core/store/src/test/java/uk/gov/gchq/gaffer/store/StoreTest.java
@@ -83,7 +83,6 @@ import uk.gov.gchq.gaffer.operation.impl.generate.GenerateElements;
 import uk.gov.gchq.gaffer.operation.impl.generate.GenerateObjects;
 import uk.gov.gchq.gaffer.operation.impl.get.GetAdjacentIds;
 import uk.gov.gchq.gaffer.operation.impl.get.GetAllElements;
-import uk.gov.gchq.gaffer.operation.impl.get.GetAsElementsFromEndpoint;
 import uk.gov.gchq.gaffer.operation.impl.get.GetElements;
 import uk.gov.gchq.gaffer.operation.impl.job.GetAllJobDetails;
 import uk.gov.gchq.gaffer.operation.impl.job.GetJobDetails;
@@ -547,7 +546,6 @@ public class StoreTest {
                 ToSingletonList.class,
                 ForEach.class,
                 Reduce.class,
-                GetAsElementsFromEndpoint.class,
 
                 // Function
                 Filter.class,
@@ -656,7 +654,6 @@ public class StoreTest {
                 ToSingletonList.class,
                 ForEach.class,
                 Reduce.class,
-                GetAsElementsFromEndpoint.class,
 
                 // Function
                 Filter.class,
